[meetup] read/service api 분리 (토큰 필요 유무로 분리) - gateway에서 토큰 인증 필터 적용하기 위해
