+++ devsta-meetup/src/main/java/com/moondy/devstameetup/controller/MeetUpSvcController.java
@@ -9,10 +9,8 @@ import com.moondy.devstameetup.domain.document.MeetUpCategory;
 import com.moondy.devstameetup.domain.dto.*;
 import com.moondy.devstameetup.service.MeetUpService;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
-import org.springframework.hateoas.Link;
 import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.util.List;
@@ -24,15 +22,15 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @RestController
 @Slf4j
-@RequestMapping("/meetup")
-public class MeetUpController {
+@RequestMapping("/meetup/service")
+public class MeetUpSvcController {
     private final MeetUpService meetUpService;
     private static final String RESULT = "result";
     private static final String CATEGORY_ALL = "ALL";
     private final MeetUpSummaryAssembler meetUpSummaryAssembler;
 
 
-    public MeetUpController(MeetUpService meetUpService, MeetUpSummaryAssembler meetUpSummaryAssembler) {
+    public MeetUpSvcController(MeetUpService meetUpService, MeetUpSummaryAssembler meetUpSummaryAssembler) {
         this.meetUpService = meetUpService;
         this.meetUpSummaryAssembler = meetUpSummaryAssembler;
     }
@@ -45,46 +43,21 @@ public class MeetUpController {
         return new CommonResponse(CommonCode.SUCCESS, Map.of(RESULT, newMeetUp.toDto()));
     }
 
-    @GetMapping("/getCategories")
-    public CommonResponse getCategoryList() {
-        List<MeetUpCategory> categoryList = meetUpService.getCategory();
-        return new CommonResponse(CommonCode.SUCCESS, Map.of(RESULT, categoryList));
-    }
 
-    @GetMapping("/getMeetUpSummaries/{category}")
-    public CollectionModel<EntityModel<MeetUpSummaryDto>> getMeetSummaries(@PathVariable("category") String category, @RequestParam int page, @RequestParam int size) {
-        // FE만 요청할 것이므로 카테고리가 정확하게 들어온다고 가정하고 코드를 작성. READ라 많이 호출될 것이기 때문에 DB 조회 최소화
-        // 대문자로 변환
-        String categoryUpper = category.toUpperCase();
-        List<MeetUp> meetUpList;
-        if (categoryUpper.equals(CATEGORY_ALL)) {
-            meetUpList = meetUpService.getRecentMeetUpSummary(page, size);
-        } else {
-            meetUpList = meetUpService.getRecentMeetUpSummaryByCategory(page, size, categoryUpper);
-        }
-        return meetUpSummaryAssembler.toCollection(meetUpList,
-                linkTo(methodOn(MeetUpController.class).getMeetSummaries(category, page + 1, size)).withRel("next"));
-    }
 
     @GetMapping("/getMyMeetUp")
     public CollectionModel<EntityModel<MeetUpSummaryDto>> getMyMeetUp(@RequestHeader("userId") String userId, @RequestParam int page, @RequestParam int size) {
         //page는 0번부터 시작
         List<MeetUp> meetUpList = meetUpService.getRecentMyMeetUpSummary(userId, page, size);
         return meetUpSummaryAssembler.toCollection(meetUpList,
-                linkTo(methodOn(MeetUpController.class).getMyMeetUp(userId, page + 1, size)).withRel("next"));
+                linkTo(methodOn(MeetUpSvcController.class).getMyMeetUp(userId, page + 1, size)).withRel("next"));
     }
 
     @GetMapping("/getJoinedMeetUp")
     public CollectionModel<EntityModel<MeetUpSummaryDto>> getJoinedMeetUp(@RequestHeader("userId") String userId,@RequestParam int page, @RequestParam int size) {
         List<MeetUp> meetUpList = meetUpService.getJoinedMeetUpSummary(userId, page, size);
         return meetUpSummaryAssembler.toCollection(meetUpList,
-                linkTo(methodOn(MeetUpController.class).getJoinedMeetUp(userId, page +1, size)).withRel("next"));
-    }
-
-    @GetMapping("/getOneMeetUp")
-    public CommonResponse getOneMeetUp(@RequestParam String meetUpId) {
-        MeetUp meetUp = meetUpService.getOneMeetUp(meetUpId);
-        return new CommonResponse(CommonCode.SUCCESS, Map.of(RESULT, meetUp.toDto()));
+                linkTo(methodOn(MeetUpSvcController.class).getJoinedMeetUp(userId, page +1, size)).withRel("next"));
     }
 
     @GetMapping("/getMeetUpStatus")
